moved game objects array
